Provide optimized option for user->posts subquery
In some cases (specifically when there are many more posts than users), doing a sub-query that is a limited select based off author is a lot faster than an `IN` query. Because this is a hint that really is specific to each site, I've put the optimization behind a filter flag.